fix: better deserialize and macros
